fix subpath and rendering
