+++ web/src/user/UserInterface.ts
@@ -1,12 +1,7 @@
 import { DEFAULT_CONFIG } from "@goauthentik/common/api/config";
-import {
-    EVENT_API_DRAWER_TOGGLE,
-    EVENT_NOTIFICATION_DRAWER_TOGGLE,
-    EVENT_WS_MESSAGE,
-} from "@goauthentik/common/constants";
-import { globalAK } from "@goauthentik/common/global";
+import { EVENT_API_DRAWER_TOGGLE, EVENT_NOTIFICATION_DRAWER_TOGGLE, EVENT_WS_MESSAGE } from "@goauthentik/common/constants";
 import { configureSentry } from "@goauthentik/common/sentry";
-import { UIConfig, UserDisplay } from "@goauthentik/common/ui/config";
+import { UIConfig } from "@goauthentik/common/ui/config";
 import { me } from "@goauthentik/common/users";
 import { WebsocketClient } from "@goauthentik/common/ws";
 import { AKElement } from "@goauthentik/elements/Base";
@@ -15,6 +10,7 @@ import "@goauthentik/elements/ak-locale-context";
 import "@goauthentik/elements/buttons/ActionButton";
 import "@goauthentik/elements/enterprise/EnterpriseStatusBanner";
 import "@goauthentik/elements/messages/MessageContainer";
+import "@goauthentik/components/ak-nav-buttons";
 import "@goauthentik/elements/notifications/APIDrawer";
 import "@goauthentik/elements/notifications/NotificationDrawer";
 import { getURLParam, updateURLParams } from "@goauthentik/elements/router/RouteMatch";
@@ -25,11 +21,14 @@ import "@goauthentik/elements/sidebar/SidebarItem";
 import { themeImage } from "@goauthentik/elements/utils/images";
 import { ROUTES } from "@goauthentik/user/Routes";
 import "@patternfly/elements/pf-tooltip/pf-tooltip.js";
-import { match } from "ts-pattern";
+
+
 
 import { msg } from "@lit/localize";
 import { css, html, nothing } from "lit";
 import { customElement, property, state } from "lit/decorators.js";
+
+
 
 import PFAvatar from "@patternfly/patternfly/components/Avatar/avatar.css";
 import PFBrand from "@patternfly/patternfly/components/Brand/brand.css";
@@ -41,7 +40,10 @@ import PFPage from "@patternfly/patternfly/components/Page/page.css";
 import PFBase from "@patternfly/patternfly/patternfly-base.css";
 import PFDisplay from "@patternfly/patternfly/utilities/Display/display.css";
 
-import { CoreApi, CurrentBrand, EventsApi, SessionUser } from "@goauthentik/api";
+
+
+import { CurrentBrand, EventsApi, SessionUser } from "@goauthentik/api";
+
 
 const customStyles = css`
     .pf-c-page__main,
@@ -152,14 +154,6 @@ class UserInterfacePresentation extends AKElement {
     @property({ type: Object })
     brand!: CurrentBrand;
 
-    get userDisplayName() {
-        return match<UserDisplay, string>(this.uiConfig.navbar.userDisplay)
-            .with(UserDisplay.username, () => this.me.user.username)
-            .with(UserDisplay.name, () => this.me.user.name)
-            .with(UserDisplay.email, () => this.me.user.email || "")
-            .otherwise(() => this.me.user.username);
-    }
-
     get canAccessAdmin() {
         return (
             this.me.user.isSuperuser ||
@@ -172,6 +166,19 @@ class UserInterfacePresentation extends AKElement {
         return Boolean(this.uiConfig && this.me && this.brand);
     }
 
+    renderAdminInterfaceLink() {
+        if (!this.canAccessAdmin) {
+            return nothing;
+        }
+
+        return html`<a
+            class="pf-c-button pf-m-secondary pf-m-small pf-u-display-none pf-u-display-block-on-md"
+            href="/if/admin/"
+            slot="extra"
+        >
+            ${msg("Admin interface")}
+        </a>`;
+    }
     render() {
         // The `!` in the field definitions above only re-assure typescript and eslint that the
         // values *should* be available, not that they *are*. Thus this contract check; it asserts
@@ -181,7 +188,7 @@ class UserInterfacePresentation extends AKElement {
             throw new Error("ak-interface-user-presentation misused; no valid values passed");
         }
 
-        return html` <ak-locale-context>
+        return html`<ak-locale-context>
             <ak-enterprise-status interface="user"></ak-enterprise-status>
             <div class="pf-c-page">
                 <div class="background-wrapper" style="${this.uiConfig.theme.background}">
@@ -199,39 +206,7 @@ class UserInterfacePresentation extends AKElement {
                             />
                         </a>
                     </div>
-                    <div class="pf-c-page__header-tools">
-                        <div class="pf-c-page__header-tools-group">
-                            ${this.renderApiDrawerTrigger()}
-                            <!-- -->
-                            ${this.renderNotificationDrawerTrigger()}
-                            <!-- -->
-                            ${this.renderSettings()}
-                            <div class="pf-c-page__header-tools-item">
-                                <a
-                                    href="${globalAK().api.base}flows/-/default/invalidation/"
-                                    class="pf-c-button pf-m-plain"
-                                >
-                                    <pf-tooltip position="top" content=${msg("Sign out")}>
-                                        <i class="fas fa-sign-out-alt" aria-hidden="true"></i>
-                                    </pf-tooltip>
-                                </a>
-                            </div>
-                            ${this.renderAdminInterfaceLink()}
-                        </div>
-                        ${this.renderImpersonation()}
-                        <div class="pf-c-page__header-tools-group">
-                            <div
-                                class="pf-c-page__header-tools-item pf-m-hidden pf-m-visible-on-md"
-                            >
-                                ${this.userDisplayName}
-                            </div>
-                        </div>
-                        <img
-                            class="pf-c-avatar"
-                            src=${this.me.user.avatar}
-                            alt="${msg("Avatar image")}"
-                        />
-                    </div>
+                    <ak-nav-buttons .uiConfig=${this.uiConfig} .me=${this.me}>${this.renderAdminInterfaceLink()}</ak-nav-buttons>
                 </header>
                 <div class="pf-c-page__drawer">
                     <div
@@ -273,108 +248,6 @@ class UserInterfacePresentation extends AKElement {
                 </div>
             </div>
         </ak-locale-context>`;
-    }
-
-    renderApiDrawerTrigger() {
-        if (!this.uiConfig.enabledFeatures.apiDrawer) {
-            return nothing;
-        }
-
-        const onClick = (ev: Event) => {
-            ev.stopPropagation();
-            this.dispatchEvent(
-                new Event(EVENT_API_DRAWER_TOGGLE, { bubbles: true, composed: true }),
-            );
-        };
-
-        return html`<div class="pf-c-page__header-tools-item pf-m-hidden pf-m-visible-on-lg">
-            <button class="pf-c-button pf-m-plain" type="button" @click=${onClick}>
-                <pf-tooltip position="top" content=${msg("Open API drawer")}>
-                    <i class="fas fa-code" aria-hidden="true"></i>
-                </pf-tooltip>
-            </button>
-        </div>`;
-    }
-
-    renderNotificationDrawerTrigger() {
-        if (!this.uiConfig.enabledFeatures.notificationDrawer) {
-            return nothing;
-        }
-
-        const onClick = (ev: Event) => {
-            ev.stopPropagation();
-            this.dispatchEvent(
-                new Event(EVENT_NOTIFICATION_DRAWER_TOGGLE, { bubbles: true, composed: true }),
-            );
-        };
-
-        return html`<div class="pf-c-page__header-tools-item pf-m-hidden pf-m-visible-on-lg">
-            <button
-                class="pf-c-button pf-m-plain"
-                type="button"
-                aria-label="${msg("Unread notifications")}"
-                @click=${onClick}
-            >
-                <span
-                    class="pf-c-notification-badge ${this.notificationsCount > 0
-                        ? "pf-m-unread"
-                        : ""}"
-                >
-                    <pf-tooltip position="top" content=${msg("Open Notification drawer")}>
-                        <i class="fas fa-bell" aria-hidden="true"></i>
-                    </pf-tooltip>
-                    <span class="pf-c-notification-badge__count">${this.notificationsCount}</span>
-                </span>
-            </button>
-        </div> `;
-    }
-
-    renderSettings() {
-        if (!this.uiConfig.enabledFeatures.settings) {
-            return nothing;
-        }
-
-        return html` <div class="pf-c-page__header-tools-item">
-            <a class="pf-c-button pf-m-plain" type="button" href="#/settings">
-                <pf-tooltip position="top" content=${msg("Settings")}>
-                    <i class="fas fa-cog" aria-hidden="true"></i>
-                </pf-tooltip>
-            </a>
-        </div>`;
-    }
-
-    renderAdminInterfaceLink() {
-        if (!this.canAccessAdmin) {
-            return nothing;
-        }
-
-        return html`<a
-            class="pf-c-button pf-m-secondary pf-m-small pf-u-display-none pf-u-display-block-on-md"
-            href="${globalAK().api.base}if/admin/"
-        >
-            ${msg("Admin interface")}
-        </a>`;
-    }
-
-    renderImpersonation() {
-        if (!this.me.original) {
-            return nothing;
-        }
-
-        const onClick = () => {
-            return new CoreApi(DEFAULT_CONFIG).coreUsersImpersonateEndRetrieve().then(() => {
-                window.location.reload();
-            });
-        };
-
-        return html`&nbsp;
-            <div class="pf-c-page__header-tools">
-                <div class="pf-c-page__header-tools-group">
-                    <ak-action-button class="pf-m-warning pf-m-small" .apiRequest=${onClick}>
-                        ${msg("Stop impersonation")}
-                    </ak-action-button>
-                </div>
-            </div>`;
     }
 }
